remove separate offline map list fragment
