+++ app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/infrastructure/MapPickers.kt
@@ -6,8 +6,7 @@ import com.kylecorry.trail_sense.shared.grouping.lists.GroupListManager
 import com.kylecorry.trail_sense.shared.grouping.picker.GroupablePickers
 import com.kylecorry.trail_sense.tools.offline_maps.domain.groups.MapGroup
 import com.kylecorry.trail_sense.tools.offline_maps.infrastructure.groups.MapGroupLoader
-import com.kylecorry.trail_sense.tools.offline_maps.infrastructure.MapService
-import com.kylecorry.trail_sense.tools.offline_maps.ui.photo_maps.mappers.MapGroupMapper
+import com.kylecorry.trail_sense.tools.offline_maps.ui.mappers.MapGroupMapper
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlin.coroutines.resume
