+++ app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/infrastructure/vector_maps/MapFileTypeUtils.kt
@@ -18,7 +18,7 @@ object MapFileTypeUtils {
         }
     }
 
-    suspend fun getExtension(type: OfflineMapFileType): String {
+    fun getExtension(type: OfflineMapFileType): String {
         return when (type) {
             OfflineMapFileType.Mapsforge -> "map"
         }
