+++ app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/ui/commands/ResizeMapCommand.kt
@@ -1,4 +1,4 @@
-package com.kylecorry.trail_sense.tools.offline_maps.ui.photo_maps.commands
+package com.kylecorry.trail_sense.tools.offline_maps.ui.commands
 
 import android.content.Context
 import com.kylecorry.andromeda.alerts.loading.ILoadingIndicator
