+++ app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/ui/commands/ToggleVisibilityMapCommand.kt
@@ -1,4 +1,4 @@
-package com.kylecorry.trail_sense.tools.offline_maps.ui.photo_maps.commands
+package com.kylecorry.trail_sense.tools.offline_maps.ui.commands
 
 import com.kylecorry.trail_sense.shared.commands.generic.CoroutineCommand
 import com.kylecorry.trail_sense.tools.offline_maps.domain.IMap
