+++ app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/ui/mappers/MapGroupAction.kt
@@ -1,4 +1,4 @@
-package com.kylecorry.trail_sense.tools.offline_maps.ui.photo_maps.mappers
+package com.kylecorry.trail_sense.tools.offline_maps.ui.mappers
 
 enum class MapGroupAction {
     View,
