+++ app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/ui/mappers/OfflineMapFileAction.kt
@@ -1,4 +1,4 @@
-package com.kylecorry.trail_sense.tools.offline_maps.ui.vector_maps
+package com.kylecorry.trail_sense.tools.offline_maps.ui.mappers
 
 enum class OfflineMapFileAction {
     View,
