--- app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/ui/vector_maps/OfflineMapFileGroupAction.kt
@@ -1,10 +1,0 @@
-package com.kylecorry.trail_sense.tools.offline_maps.ui.vector_maps
-
-enum class OfflineMapFileGroupAction {
-    View,
-    Rename,
-    Delete,
-    Move,
-    ShowAll,
-    HideAll
-}
