--- app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/ui/vector_maps/commands/OfflineMapCleanupCommand.kt
@@ -1,37 +1,0 @@
-package com.kylecorry.trail_sense.tools.offline_maps.ui.vector_maps.commands
-
-import com.kylecorry.andromeda.core.coroutines.onIO
-import com.kylecorry.trail_sense.main.getAppService
-import com.kylecorry.trail_sense.shared.commands.CoroutineValueCommand
-import com.kylecorry.trail_sense.shared.io.FileSubsystem
-import com.kylecorry.trail_sense.tools.offline_maps.infrastructure.persistence.MapRepo
-
-class OfflineMapCleanupCommand : CoroutineValueCommand<Boolean> {
-
-    private val repo = getAppService<MapRepo>()
-    private val files = getAppService<FileSubsystem>()
-
-    override suspend fun execute(): Boolean = onIO {
-        val maps = repo.getVectorMaps()
-        val allFiles = files.list(OFFLINE_MAPS_DIRECTORY).map { "$OFFLINE_MAPS_DIRECTORY/${it.name}" }
-
-        // Delete files without a map
-        val mapFiles = maps.map { it.path }
-        val orphanedFiles = allFiles.filter { !mapFiles.contains(it) }
-        orphanedFiles.forEach {
-            files.delete(it)
-        }
-
-        // Delete maps without a file
-        val toDelete = maps.filter { !allFiles.contains(it.path) }
-        toDelete.forEach {
-            repo.delete(it)
-        }
-
-        toDelete.isNotEmpty()
-    }
-
-    companion object {
-        private const val OFFLINE_MAPS_DIRECTORY = "offline_maps"
-    }
-}
