+++ app/src/main/res/layout/fragment_offline_map_list.xml
@@ -6,15 +6,17 @@
     android:layout_height="match_parent">
 
     <com.kylecorry.andromeda.views.toolbar.Toolbar
-        android:id="@+id/title"
-        android:layout_width="0dp"
+        android:id="@+id/map_list_title"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:flattenButtons="true"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
+        app:leftButtonIcon="@drawable/ic_help"
+        app:rightButtonIcon="@drawable/ic_menu_dots"
         app:showSubtitle="false"
-        app:title="@string/vector_maps" />
+        app:title="@string/photo_maps" />
 
     <com.kylecorry.trail_sense.shared.views.SearchView
         android:id="@+id/searchbox"
@@ -24,30 +26,30 @@
         android:layout_marginEnd="16dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/title" />
+        app:layout_constraintTop_toBottomOf="@id/map_list_title" />
 
     <com.kylecorry.andromeda.views.list.AndromedaListView
-        android:id="@+id/list"
+        android:id="@+id/map_list"
         style="@style/AppTheme.ListViewWithFab"
-        android:layout_width="0dp"
+        android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_marginTop="16dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/searchbox" />
+        app:layout_constraintTop_toBottomOf="@id/searchbox" />
 
     <TextView
-        android:id="@+id/empty_text"
+        android:id="@+id/map_empty_text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/no_offline_maps"
+        android:text="@string/no_maps"
         android:textAppearance="@style/TextAppearance.Material3.BodyLarge"
         android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintBottom_toBottomOf="@id/map_list"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/list"
+        app:layout_constraintTop_toTopOf="@+id/map_list"
         tools:visibility="visible" />
 
     <ImageView
@@ -71,7 +73,7 @@
         android:visibility="gone"
         app:layout_constraintBottom_toTopOf="@id/add_btn"
         app:layout_constraintEnd_toEndOf="@id/add_btn"
-        app:menu="@menu/add_offline_map_menu"
+        app:menu="@menu/add_map_menu"
         tools:visibility="visible" />
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
