--- app/src/main/res/layout/fragment_photo_map_list.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <com.kylecorry.andromeda.views.toolbar.Toolbar
-        android:id="@+id/map_list_title"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:flattenButtons="true"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:leftButtonIcon="@drawable/ic_help"
-        app:rightButtonIcon="@drawable/ic_menu_dots"
-        app:showSubtitle="false"
-        app:title="@string/photo_maps" />
-
-    <com.kylecorry.trail_sense.shared.views.SearchView
-        android:id="@+id/searchbox"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginEnd="16dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/map_list_title" />
-
-    <com.kylecorry.andromeda.views.list.AndromedaListView
-        android:id="@+id/map_list"
-        style="@style/AppTheme.ListViewWithFab"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/searchbox" />
-
-    <TextView
-        android:id="@+id/map_empty_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/no_maps"
-        android:textAppearance="@style/TextAppearance.Material3.BodyLarge"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="@id/map_list"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/map_list"
-        tools:visibility="visible" />
-
-    <ImageView
-        android:id="@+id/overlay_mask"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:alpha="0.86"
-        android:background="?android:colorBackground"
-        android:clickable="true"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <com.kylecorry.trail_sense.shared.views.FloatingActionButtonMenu
-        android:id="@+id/add_menu"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="16dp"
-        android:visibility="gone"
-        app:layout_constraintBottom_toTopOf="@id/add_btn"
-        app:layout_constraintEnd_toEndOf="@id/add_btn"
-        app:menu="@menu/add_map_menu"
-        tools:visibility="visible" />
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/add_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:layout_marginBottom="@dimen/default_bottom_margin"
-        android:clickable="true"
-        android:focusable="true"
-        android:src="@drawable/ic_add"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
